"points" with float components (aka dotz::vec2)
currently, mouse coords and window size is used only as floats and under
certain math operations (aspect, coord transformation, etc). Therefore
it is easier if casein provides its "points" wrapped in a dotz::vec2.

+++ CASView.mm
@@ -55,8 +55,8 @@ extern casein::point * casein_window_size;
 
 - (void)sendResizeEvent {
   *casein_window_size = {
-    static_cast<int>(self.frame.size.width),
-    static_cast<int>(self.frame.size.height),
+    static_cast<float>(self.frame.size.width),
+    static_cast<float>(self.frame.size.height),
   };
   *casein_screen_scale_factor = [self backingScaleFactor];
   // [self inLiveResize] == YES,
